@@ -1,75 +1,100 @@
 {
-  "name": "@danky/web",
-  "version": "0.0.1",
+  "name": "web",
+  "version": "0.1.0",
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "kill-port 3000 && next dev",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint . --report-unused-disable-directives",
-    "lint:fix": "eslint . --fix",
-    "type-check": "tsc --noEmit",
-    "clean": "rimraf .next .turbo node_modules",
-    "ui:add": "pnpm dlx shadcn-ui@latest add",
-    "ui:diff": "pnpm dlx shadcn-ui@latest diff"
+    "lint": "eslint .",
+    "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "@danky/db": "workspace:*",
+    "@anthropic-ai/sdk": "^0.33.1",
     "@danky/logger": "workspace:*",
-    "@danky/mcp": "workspace:*",
+    "@danky/qdrant": "workspace:*",
+    "@danky/redis": "workspace:*",
     "@danky/schema": "workspace:*",
     "@danky/ui": "workspace:*",
-    "@h1deya/mcp-langchain-tools": "^0.1.1",
-    "@langchain/anthropic": "^0.3.11",
-    "@langchain/community": "^0.3.22",
-    "@langchain/core": "^0.3.27",
-    "@langchain/groq": "^0.1.2",
-    "@langchain/langgraph": "^0.2.36",
-    "@langchain/openai": "^0.3.16",
-    "@modelcontextprotocol/sdk": "^1.1.0",
-    "@radix-ui/react-slot": "^1.0.2",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.0",
-    "drizzle-orm": "^0.30.2",
-    "lucide-react": "^0.323.0",
+    "@hookform/resolvers": "3.10.0",
+    "@radix-ui/react-accordion": "^1.2.2",
+    "@radix-ui/react-alert-dialog": "^1.1.4",
+    "@radix-ui/react-aspect-ratio": "1.1.1",
+    "@radix-ui/react-avatar": "^1.1.2",
+    "@radix-ui/react-checkbox": "1.1.3",
+    "@radix-ui/react-collapsible": "1.1.2",
+    "@radix-ui/react-context-menu": "2.2.4",
+    "@radix-ui/react-dialog": "1.1.4",
+    "@radix-ui/react-dropdown-menu": "2.1.4",
+    "@radix-ui/react-hover-card": "1.1.4",
+    "@radix-ui/react-label": "2.1.1",
+    "@radix-ui/react-menubar": "1.1.4",
+    "@radix-ui/react-navigation-menu": "1.2.3",
+    "@radix-ui/react-popover": "1.1.4",
+    "@radix-ui/react-progress": "1.1.1",
+    "@radix-ui/react-radio-group": "1.2.2",
+    "@radix-ui/react-scroll-area": "1.2.2",
+    "@radix-ui/react-select": "2.1.4",
+    "@radix-ui/react-separator": "1.1.1",
+    "@radix-ui/react-slider": "1.2.2",
+    "@radix-ui/react-slot": "^1.1.1",
+    "@radix-ui/react-switch": "1.1.2",
+    "@radix-ui/react-tabs": "1.1.2",
+    "@radix-ui/react-toast": "1.2.4",
+    "@radix-ui/react-toggle": "1.1.1",
+    "@radix-ui/react-tooltip": "1.1.6",
+    "clsx": "^2.1.1",
+    "cmdk": "^1.0.4",
+    "date-fns": "4.1.0",
+    "lucide-react": "^0.471.0",
     "next": "^15.1.4",
-    "next-themes": "^0.2.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-markdown": "^9.0.1",
-    "rehype-katex": "^7.0.0",
-    "rehype-raw": "^7.0.0",
-    "remark-gfm": "^4.0.0",
-    "remark-math": "^6.0.0",
-    "shiki": "^1.1.7",
-    "react-syntax-highlighter": "^15.5.0",
-    "tailwind-merge": "^2.2.1",
-    "tailwindcss-animate": "^1.0.7",
-    "zustand": "^4.5.2"
+    "next-themes": "^0.4.4",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-hook-form": "7.54.2",
+    "sonner": "^1.7.1",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@danky/eslint-config": "workspace:*",
     "@danky/typescript-config": "workspace:*",
-    "@eslint/eslintrc": "3.2.0",
-    "@eslint/js": "^9.0.0",
-    "@next/eslint-plugin-next": "^15.1.4",
-    "@types/node": "^20",
-    "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.0",
-    "@types/react-syntax-highlighter": "^15.5.11",
-    "@typescript-eslint/eslint-plugin": "^8.19.1",
-    "@typescript-eslint/parser": "^8.19.1",
-    "autoprefixer": "^10.0.1",
+    "@types/node": "^22.10.5",
+    "@types/react": "^19.0.4",
+    "@types/react-dom": "^19.0.2",
+    "autoprefixer": "^10.4.20",
+    "class-variance-authority": "^0.7.0",
     "eslint": "^9.17.0",
-    "eslint-config-next": "^15.1.4",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "postcss": "^8",
-    "rimraf": "^5.0.5",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5.3.3"
+    "postcss": "^8.4.49",
+    "tailwind-merge": "^2.6.0",
+    "tailwindcss": "^3.4.17",
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "^5.4.5",
+    "@types/d3-array": "^3.2.1",
+    "@types/d3-color": "^3.1.3",
+    "@types/d3-ease": "^3.0.2",
+    "@types/d3-interpolate": "^3.0.4",
+    "@types/d3-path": "^3.0.2",
+    "@types/d3-scale": "^4.0.8",
+    "@types/d3-shape": "^3.1.6",
+    "@types/d3-time": "^3.0.3",
+    "@types/d3-timer": "^3.0.2",
+    "@types/debug": "^4.1.12",
+    "@types/estree": "^1.0.5",
+    "@types/glob": "^8.1.0",
+    "@types/inquirer": "^9.0.7",
+    "@types/json-schema": "^7.0.15",
+    "@types/minimatch": "^5.1.2",
+    "@types/ms": "^0.7.34",
+    "@types/node-fetch": "^2.6.11",
+    "@types/pg": "^8.11.0",
+    "@types/retry": "^0.12.5",
+    "@types/through": "^0.0.33",
+    "@types/tinycolor2": "^1.4.6",
+    "@types/tough-cookie": "^4.0.5",
+    "@types/uuid": "^9.0.8",
+    "@types/ws": "^8.5.10",
+    "@types/yargs": "^17.0.32",
+    "@types/yargs-parser": "^21.0.3"
   }
 }
