@@ -1,6 +1,6 @@
 {
   "name": "@danky/db",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "private": true,
   "type": "module",
   "main": "./dist/index.js",
@@ -11,27 +11,31 @@
   ],
   "scripts": {
     "build": "tsup",
-    "clean": "rimraf .turbo node_modules dist",
-    "generate": "drizzle-kit generate:pg",
-    "migrate": "tsx src/migrate.ts",
+    "dev": "tsup --watch",
+    "type-check": "tsc --noEmit",
+    "lint": "eslint .",
+    "clean": "rimraf dist",
+    "migrate:generate": "drizzle-kit generate:pg",
+    "migrate:push": "tsx src/migrate.ts",
     "studio": "drizzle-kit studio"
   },
   "dependencies": {
     "@danky/logger": "workspace:*",
     "@paralleldrive/cuid2": "^2.2.2",
-    "dotenv": "^16.4.5",
-    "pg": "^8.11.3",
-    "postgres": "^3.4.3"
+    "drizzle-orm": "^0.38.3",
+    "pg": "^8.13.1",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
+    "@danky/eslint-config": "workspace:*",
     "@danky/typescript-config": "workspace:*",
-    "@types/node": "^20.17.12",
+    "@types/node": "^22.10.5",
     "@types/pg": "^8.11.10",
-    "drizzle-kit": "^0.20.14",
-    "drizzle-orm": "^0.29.5",
-    "rimraf": "^5.0.5",
+    "drizzle-kit": "^0.30.1",
+    "drizzle-zod": "^0.6.1",
+    "tsx": "^4.19.0",
     "tsup": "^8.0.2",
-    "tsx": "^4.7.1",
-    "typescript": "^5.3.3"
+    "typescript": "^5.4.5",
+    "rimraf": "^6.0.1"
   }
 }
