@@ -15,20 +15,22 @@
     "next.js"
   ],
   "dependencies": {
-    "@eslint/eslintrc": "^3.0.0",
-    "@eslint/js": "^9.0.0",
-    "@typescript-eslint/eslint-plugin": "^7.0.0",
-    "@typescript-eslint/parser": "^7.0.0",
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.18.0",
+    "@typescript-eslint/eslint-plugin": "^8.19.0",
+    "@typescript-eslint/parser": "^8.19.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-config-turbo": "^1.11.2",
+    "eslint-config-turbo": "^2.3.3",
     "eslint-plugin-only-warn": "^1.1.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-turbo": "^1.13.4",
-    "globals": "^13.24.0"
+    "eslint-plugin-turbo": "^2.3.3"
   },
   "peerDependencies": {
-    "eslint": "^9.0.0",
-    "typescript": ">=4.7.4"
+    "eslint": "^9.18.0",
+    "typescript": "^5.4.5",
+    "eslint-plugin-react": "^7.37.3",
+    "eslint-plugin-react-hooks": "^5.1.0"
+  },
+  "devDependencies": {
+    "globals": "^15.14.0"
   }
 }
