@@ -1,6 +1,6 @@
 {
   "name": "@danky/logger",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "private": true,
   "type": "module",
   "main": "./dist/index.js",
@@ -11,23 +11,22 @@
   ],
   "scripts": {
     "build": "tsup",
-    "clean": "rimraf .turbo node_modules dist",
     "dev": "tsup --watch",
+    "type-check": "tsc --noEmit",
     "lint": "eslint .",
-    "type-check": "tsc --noEmit"
+    "clean": "rimraf dist"
   },
   "dependencies": {
-    "pino": "^8.17.2",
-    "pino-pretty": "^10.3.1"
+    "pino": "^9.6.0",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@danky/eslint-config": "workspace:*",
     "@danky/typescript-config": "workspace:*",
-    "@types/node": "^20.10.6",
-    "@types/pino": "7.0.5",
-    "eslint": "^8.56.0",
-    "rimraf": "^5.0.5",
-    "tsup": "^8.0.1",
-    "typescript": "^5.3.3"
+    "@types/node": "^22.10.5",
+    "pino-pretty": "^13.0.0",
+    "tsup": "^8.0.2",
+    "typescript": "^5.4.5",
+    "rimraf": "^6.0.1"
   }
 }
