@@ -7,32 +7,17 @@ import { withUser } from '@/lib/session';
 import { withErrorHandling } from '@/lib/errors';
 
 async function createMessageHandler(input: z.infer<typeof createMessageRequestSchema>) {
-  return withUser(async (user) => {
+  return withUser(async user => {
     // Validate input
     const validatedInput = createMessageRequestSchema.parse(input);
 
     // Create message in database
-    const message = await createMessage(validatedInput, user.id);
-
-    // Transform message to match our schema
-    const transformedMessage = {
-      id: message.id,
-      role: message.role,
-      content: [{
-        type: message.contentType,
-        content: message.content,
-        language: message.language,
-      }],
-      sessionId: message.conversationId,
-      metadata: message.metadata,
-      createdAt: message.createdAt,
-      updatedAt: message.updatedAt,
-    };
+    const result = await createMessage(validatedInput, user.id);
 
     // Return success response
     return createMessageResponseSchema.parse({
       success: true,
-      data: transformedMessage,
+      data: result[0],
     });
   });
 }
@@ -42,4 +27,4 @@ export const createChatMessage = withErrorHandling(
   createMessageHandler,
   createMessageResponseSchema,
   'createMessage'
-); 
+);
