@@ -1,64 +1,77 @@
-import { NextResponse } from 'next/server'
-import { createLogger } from '@danky/mcp'
-import { MCPService } from '@danky/mcp'
-import path from 'path'
+import { NextRequest, NextResponse } from 'next/server';
+import {
+  createSession,
+  getSessions,
+  updateSession,
+  deleteSession,
+  batchDeleteSessions,
+} from '@/lib/db/chat';
+import { MODEL_IDS } from '@danky/schema';
 
-const logger = createLogger({ level: 'debug' })
-const mcpService = new MCPService({
-  configPath: path.join(process.cwd(), '../../mcp-config.json5'),
-  logger
-})
+export async function POST(request: NextRequest) {
+  const searchParams = request.nextUrl.searchParams;
+  const modelId = searchParams.get('modelId');
 
-export async function POST(request: Request) {
-  try {
-    const { message } = await request.json()
-    
-    if (!message || typeof message !== 'string') {
-      logger.error('Invalid message format:', { message })
-      return NextResponse.json(
-        { error: 'Invalid message format' },
-        { status: 400 }
-      )
-    }
-
-    logger.debug('Processing message:', { message })
-    const response = await mcpService.processMessage(message)
-    logger.debug('Response:', { response })
-    
-    return NextResponse.json({ response })
-  } catch (error) {
-    logger.error('Error processing message:', {
-      error: error instanceof Error ? {
-        name: error.name,
-        message: error.message,
-        stack: error.stack
-      } : String(error)
-    })
-    return NextResponse.json(
-      { error: error instanceof Error ? error.message : String(error) },
-      { status: 500 }
-    )
+  if (!modelId || !MODEL_IDS.includes(modelId as (typeof MODEL_IDS)[number])) {
+    return Response.json({ error: 'Invalid model ID' }, { status: 400 });
   }
+
+  const session = await createSession(
+    {
+      title: 'New Chat',
+      modelId: modelId as (typeof MODEL_IDS)[number],
+      description: 'Created via API',
+    },
+    'temp-user-id'
+  );
+
+  return Response.json(session);
 }
 
-export async function GET() {
-  try {
-    logger.debug('Initializing MCP service')
-    await mcpService.initialize()
-    logger.debug('MCP service initialized')
-    
-    return NextResponse.json({ status: 'ready' })
-  } catch (error) {
-    logger.error('Error initializing chat:', {
-      error: error instanceof Error ? {
-        name: error.name,
-        message: error.message,
-        stack: error.stack
-      } : String(error)
-    })
-    return NextResponse.json(
-      { error: error instanceof Error ? error.message : String(error) },
-      { status: 500 }
-    )
+export async function GET(request: NextRequest) {
+  const sessions = await getSessions(
+    {
+      limit: 10,
+      page: 1,
+      includeDeleted: false,
+      sortBy: 'createdAt',
+      sortOrder: 'desc',
+    },
+    'temp-user-id'
+  );
+  return Response.json(sessions);
+}
+
+export async function PUT(request: NextRequest) {
+  const body = await request.json();
+  const session = await updateSession(body, 'temp-user-id');
+  return Response.json(session);
+}
+
+export async function DELETE(request: NextRequest) {
+  const searchParams = request.nextUrl.searchParams;
+  const ids = searchParams.getAll('id');
+  const permanent = searchParams.get('permanent') === 'true';
+
+  if (ids.length > 1) {
+    const result = await batchDeleteSessions(
+      {
+        ids,
+        permanent,
+      },
+      'temp-user-id'
+    );
+
+    return NextResponse.json(result);
   }
+
+  const result = await deleteSession(
+    {
+      id: ids[0],
+      permanent,
+    },
+    'temp-user-id'
+  );
+
+  return NextResponse.json(result);
 }
