@@ -1,73 +1,72 @@
-"use client";
+'use client';
 
-import * as React from "react";
-import { cn } from "@/lib/utils";
-import { Input } from "../input";
-import { Separator } from "../separator";
+import * as React from 'react';
+import { cn } from '@/lib/utils';
+import { Input, Separator, type InputProps, type SeparatorProps } from '@danky/ui';
 
-export const SidebarInput = React.forwardRef<
-  React.ElementRef<typeof Input>,
-  React.ComponentPropsWithoutRef<typeof Input>
->(({ className, ...props }, ref) => {
-  return (
-    <Input
-      ref={ref}
-      data-sidebar="input"
-      className={cn(
-        "h-8 w-full bg-background shadow-none focus-visible:ring-2 focus-visible:ring-sidebar-ring",
-        className
-      )}
-      {...props}
-    />
-  );
-});
+export interface SidebarInputProps extends InputProps {}
 
-SidebarInput.displayName = "SidebarInput";
+export const SidebarInput = React.forwardRef<HTMLInputElement, SidebarInputProps>(
+  ({ className, ...props }, ref) => {
+    return (
+      <Input
+        ref={ref}
+        data-sidebar="input"
+        className={cn(
+          'bg-background focus-visible:ring-sidebar-ring h-8 w-full shadow-none focus-visible:ring-2',
+          className
+        )}
+        {...props}
+      />
+    );
+  }
+);
+SidebarInput.displayName = 'SidebarInput';
 
-export const SidebarHeader = React.forwardRef<
-  HTMLDivElement,
-  React.HTMLAttributes<HTMLDivElement>
->(({ className, ...props }, ref) => {
-  return (
-    <div
-      ref={ref}
-      data-sidebar="header"
-      className={cn("flex flex-col gap-2 p-2", className)}
-      {...props}
-    />
-  );
-});
+export interface SidebarHeaderProps extends React.HTMLAttributes<HTMLDivElement> {}
 
-SidebarHeader.displayName = "SidebarHeader";
+export const SidebarHeader = React.forwardRef<HTMLDivElement, SidebarHeaderProps>(
+  ({ className, ...props }, ref) => {
+    return (
+      <div
+        ref={ref}
+        data-sidebar="header"
+        className={cn('flex flex-col gap-2 p-2', className)}
+        {...props}
+      />
+    );
+  }
+);
+SidebarHeader.displayName = 'SidebarHeader';
 
-export const SidebarFooter = React.forwardRef<
-  HTMLDivElement,
-  React.HTMLAttributes<HTMLDivElement>
->(({ className, ...props }, ref) => {
-  return (
-    <div
-      ref={ref}
-      data-sidebar="footer"
-      className={cn("flex flex-col gap-2 p-2", className)}
-      {...props}
-    />
-  );
-});
+export interface SidebarFooterProps extends React.HTMLAttributes<HTMLDivElement> {}
 
-SidebarFooter.displayName = "SidebarFooter";
+export const SidebarFooter = React.forwardRef<HTMLDivElement, SidebarFooterProps>(
+  ({ className, ...props }, ref) => {
+    return (
+      <div
+        ref={ref}
+        data-sidebar="footer"
+        className={cn('flex flex-col gap-2 p-2', className)}
+        {...props}
+      />
+    );
+  }
+);
+SidebarFooter.displayName = 'SidebarFooter';
 
-export const SidebarSeparator = React.forwardRef<
-  React.ElementRef<typeof Separator>,
-  React.ComponentPropsWithoutRef<typeof Separator>
->(({ className, ...props }, ref) => {
-  return (
-    <Separator
-      ref={ref}
-      data-sidebar="separator"
-      className={cn("mx-2 w-auto bg-sidebar-border", className)}
-      {...props}
-    />
-  );
-});
+export interface SidebarSeparatorProps extends SeparatorProps {}
 
-SidebarSeparator.displayName = "SidebarSeparator";
+export const SidebarSeparator = React.forwardRef<HTMLHRElement, SidebarSeparatorProps>(
+  ({ className, ...props }, ref) => {
+    return (
+      <Separator
+        ref={ref}
+        data-sidebar="separator"
+        className={cn('bg-sidebar-border mx-2 w-auto', className)}
+        {...props}
+      />
+    );
+  }
+);
+SidebarSeparator.displayName = 'SidebarSeparator';
