@@ -1,60 +1,94 @@
-'use server'
+'use server';
 
-import { getMCPService, shutdownMCPService } from '../services/mcp-provider'
+import { convertMcpToLangchainTools } from '@h1deya/langchain-mcp-tools';
+import { DynamicStructuredTool } from '@langchain/core/tools';
+import { ChatOpenAI } from '@langchain/openai';
+import { initializeAgentExecutorWithOptions, AgentExecutor } from 'langchain/agents';
 
-type ServerStatus = 'connected' | 'disconnected' | 'error'
+type ServerStatus = 'connected' | 'disconnected' | 'error';
 
 interface Tool {
-  serverId: string
-  toolId: string
-  name: string
-  description: string
+  serverId: string;
+  toolId: string;
+  name: string;
+  description: string;
 }
 
 interface InitializeResult {
-  isInitialized: boolean
-  serverStatus: Record<string, ServerStatus>
-  availableTools: Tool[]
-  error: string | null
+  isInitialized: boolean;
+  serverStatus: Record<string, ServerStatus>;
+  availableTools: Tool[];
+  error: string | null;
 }
+
+let agent: AgentExecutor | null = null;
+let cleanup: (() => Promise<void>) | null = null;
 
 export async function initialize(): Promise<InitializeResult> {
   try {
-    const _service = await getMCPService()
+    const model = new ChatOpenAI({
+      modelName: 'gpt-4-turbo-preview',
+      temperature: 0,
+    });
+
+    const { tools, cleanup: cleanupFn } = await convertMcpToLangchainTools({
+      mcp: {
+        command: 'node',
+        args: ['server.js'],
+        env: {
+          MCP_SERVER_URL: process.env.MCP_SERVER_URL || 'http://localhost:3001',
+        },
+      },
+    });
+
+    cleanup = cleanupFn;
+    agent = await initializeAgentExecutorWithOptions(tools, model, {
+      agentType: 'chat-conversational-react-description',
+      verbose: true,
+    });
+
     return {
       isInitialized: true,
       serverStatus: {
-        mcp: 'connected'
+        mcp: 'connected',
       },
-      availableTools: [],
-      error: null
-    }
+      availableTools: tools.map((tool: DynamicStructuredTool) => ({
+        serverId: 'mcp',
+        toolId: tool.name,
+        name: tool.name,
+        description: tool.description,
+      })),
+      error: null,
+    };
   } catch (error) {
-    console.error('Failed to initialize MCP service:', error)
+    console.error('Failed to initialize MCP service:', error);
     return {
       isInitialized: false,
       serverStatus: {},
       availableTools: [],
-      error: error instanceof Error ? error.message : String(error)
-    }
+      error: error instanceof Error ? error.message : String(error),
+    };
   }
 }
 
 export async function processMessage(message: string) {
+  if (!agent) {
+    throw new Error('Agent not initialized');
+  }
+
   try {
-    const _service = await getMCPService()
-    return await _service.processMessage(message)
+    const result = await agent.call({ input: message });
+    return result.output;
   } catch (error) {
-    console.error('Failed to process message:', error)
-    throw error
+    console.error('Failed to process message:', error);
+    throw error;
   }
 }
 
 export async function shutdown() {
-  try {
-    await shutdownMCPService()
-  } catch (error) {
-    console.error('Failed to shutdown MCP service:', error)
-    throw error
+  if (cleanup) {
+    await cleanup();
+    cleanup = null;
   }
-} 
+  agent = null;
+}
