@@ -1,6 +1,7 @@
 import { db } from '@danky/db';
 import { handleDatabaseError } from '../errors';
 import { logger } from '@danky/logger';
+import { sql } from 'drizzle-orm';
 
 // Transaction options type
 interface TransactionOptions {
@@ -22,7 +23,7 @@ const wait = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
 // Helper to check if error is retryable
 function isRetryableError(error: unknown): boolean {
   if (!(error instanceof Error)) return false;
-  
+
   const message = error.message.toLowerCase();
   return (
     message.includes('deadlock') ||
@@ -40,21 +41,21 @@ async function executeTransaction<T>(
 ): Promise<T> {
   const { maxRetries, retryDelay, timeout } = { ...DEFAULT_OPTIONS, ...options };
   let attempt = 0;
-  
+
   while (true) {
     attempt++;
-    
+
     try {
       // Start transaction with timeout
       const result = await Promise.race([
-        db.transaction(async (tx) => {
+        db.transaction(async tx => {
           return operation(tx);
         }),
-        new Promise((_, reject) => 
+        new Promise((_, reject) =>
           setTimeout(() => reject(new Error('Transaction timeout')), timeout)
         ),
       ]);
-      
+
       return result as T;
     } catch (error) {
       logger.error({
@@ -62,11 +63,11 @@ async function executeTransaction<T>(
         attempt,
         error: error instanceof Error ? error.message : 'Unknown error',
       });
-      
+
       if (attempt >= maxRetries || !isRetryableError(error)) {
         return handleDatabaseError(error);
       }
-      
+
       // Wait before retrying
       await wait(retryDelay * attempt);
     }
@@ -78,9 +79,9 @@ export async function withReadOnlyTransaction<T>(
   operation: (tx: typeof db) => Promise<T>,
   options?: TransactionOptions
 ): Promise<T> {
-  return executeTransaction(async (tx) => {
+  return executeTransaction(async tx => {
     // Set transaction to read only
-    await tx.execute({ sql: 'SET TRANSACTION READ ONLY' });
+    await tx.execute(sql`SET TRANSACTION READ ONLY`);
     return operation(tx);
   }, options);
 }
@@ -90,9 +91,9 @@ export async function withWriteTransaction<T>(
   operation: (tx: typeof db) => Promise<T>,
   options?: TransactionOptions
 ): Promise<T> {
-  return executeTransaction(async (tx) => {
+  return executeTransaction(async tx => {
     // Set transaction isolation level
-    await tx.execute({ sql: 'SET TRANSACTION ISOLATION LEVEL SERIALIZABLE' });
+    await tx.execute(sql`SET TRANSACTION ISOLATION LEVEL SERIALIZABLE`);
     return operation(tx);
   }, options);
 }
@@ -102,10 +103,10 @@ export async function withBatchTransaction<T>(
   operations: Array<(tx: typeof db) => Promise<T>>,
   options?: TransactionOptions
 ): Promise<T[]> {
-  return executeTransaction(async (tx) => {
+  return executeTransaction(async tx => {
     const results: T[] = [];
     const errors: Error[] = [];
-    
+
     for (const operation of operations) {
       try {
         const result = await operation(tx);
@@ -119,12 +120,12 @@ export async function withBatchTransaction<T>(
         });
       }
     }
-    
+
     // If all operations failed, throw error
     if (errors.length === operations.length) {
       throw new Error('All batch operations failed');
     }
-    
+
     return results;
   }, options);
 }
@@ -135,7 +136,7 @@ let poolSize = 10;
 export function setPoolSize(size: number): void {
   poolSize = size;
   // Update pool size in database configuration
-  db.execute({ sql: `SET max_connections = ${size}` });
+  db.execute(sql`SET max_connections = ${size}`);
 }
 
 export function getPoolSize(): number {
@@ -144,4 +145,4 @@ export function getPoolSize(): number {
 
 // Export transaction types
 export type Transaction = typeof db;
-export type TransactionClient = Awaited<ReturnType<typeof db.transaction>>; 
+export type TransactionClient = Awaited<ReturnType<typeof db.transaction>>;
