@@ -1,33 +1,3 @@
-import baseConfig from "@danky/eslint-config/base";
-import tseslint from "@typescript-eslint/eslint-plugin";
+import { nextConfig } from '@danky/eslint-config';
 
-/** @type {import("eslint").Linter.FlatConfig[]} */
-export default [
-  {
-    ignores: [
-      "**/dist/**",
-      "**/.turbo/**",
-      "**/node_modules/**",
-      "**/build/**",
-    ],
-  },
-  ...baseConfig,
-  {
-    files: ["**/*.{ts,tsx}"],
-    plugins: {
-      "@typescript-eslint": tseslint,
-    },
-    rules: {
-      "@typescript-eslint/no-explicit-any": "error",
-      "@typescript-eslint/explicit-function-return-type": "off",
-      "@typescript-eslint/no-unused-vars": [
-        "warn",
-        {
-          argsIgnorePattern: "^_",
-          varsIgnorePattern: "^_",
-        },
-      ],
-      "no-console": ["warn", { allow: ["warn", "error"] }],
-    },
-  },
-];
+export default nextConfig;
