@@ -1,19 +1,13 @@
-import type { Config } from "drizzle-kit";
-import * as dotenv from "dotenv";
-import { resolve } from "path";
-
-// Load environment variables from the root .env file
-dotenv.config({ path: resolve(__dirname, "../../.env") });
-
-if (!process.env.DATABASE_URL) {
-  throw new Error("DATABASE_URL environment variable is required");
-}
+import type { Config } from 'drizzle-kit';
+import { env } from './src/env';
 
 export default {
-  schema: "./src/schema/*",
-  out: "./drizzle",
-  driver: "pg",
+  schema: './src/schema/*',
+  out: './src/migrations',
+  driver: 'pg',
   dbCredentials: {
-    connectionString: process.env.DATABASE_URL,
+    connectionString: env.DATABASE_URL,
   },
-} satisfies Config; 
+  verbose: true,
+  strict: true,
+} satisfies Config;
