@@ -1,2 +1,4 @@
-export * from "./client";
-export * from "./schema"; 
+export * from './client';
+export * from './schema';
+export * from './env';
+export type { Pool } from 'pg';
