@@ -1,33 +1,33 @@
-import { drizzle } from "drizzle-orm/node-postgres";
-import pkg from "pg";
+import { drizzle } from 'drizzle-orm/node-postgres';
+import pkg from 'pg';
 const { Pool } = pkg;
-import { createLogger } from "@danky/logger";
-import { users, sessions, conversations, messages } from "./schema";
-import * as dotenv from "dotenv";
-import { resolve, dirname } from "path";
-import { fileURLToPath } from "url";
-import { randomUUID } from "crypto";
+import { createLogger } from '@danky/logger';
+import { users } from './schema';
+import * as dotenv from 'dotenv';
+import { resolve, dirname } from 'path';
+import { fileURLToPath } from 'url';
+import { createId } from '@paralleldrive/cuid2';
 
 // Get the current file's directory
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
 // Load environment variables from the root .env file
-dotenv.config({ path: resolve(__dirname, "../../../.env") });
+dotenv.config({ path: resolve(__dirname, '../../../.env') });
 
-const logger = createLogger("db:seed");
+const logger = createLogger('db:seed');
 
 // Construct the DATABASE_URL manually
 const {
-  POSTGRES_USER = "danky",
-  POSTGRES_PASSWORD = "danky",
-  POSTGRES_DB = "danky",
-  POSTGRES_PORT = "5432",
+  POSTGRES_USER = 'danky',
+  POSTGRES_PASSWORD = 'danky',
+  POSTGRES_DB = 'danky',
+  POSTGRES_PORT = '5432',
 } = process.env;
 
 const DATABASE_URL = `postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@localhost:${POSTGRES_PORT}/${POSTGRES_DB}`;
 
-logger.info("Environment variables:", {
+logger.info('Environment variables:', {
   DATABASE_URL,
   POSTGRES_USER,
   POSTGRES_PASSWORD,
@@ -44,54 +44,31 @@ const db = drizzle(pool);
 
 // Seed data
 async function main() {
-  logger.info("Seeding database...");
-  
+  logger.info('Seeding database...');
+
   try {
     // Create test user
-    const [user] = await db.insert(users).values({
-      id: randomUUID(),
-      email: "test@example.com",
-      hashedPassword: "hashed_password_here", // In production, use proper password hashing
-      fullName: "Test User",
-      isActive: true,
-      isVerified: true,
-    }).returning();
+    const [user] = await db
+      .insert(users)
+      .values({
+        id: createId(),
+        email: 'test@example.com',
+        name: 'Test User',
+        metadata: JSON.stringify({ isActive: true, isVerified: true }),
+      })
+      .returning();
 
-    logger.info({ userId: user.id }, "Created test user");
-
-    // Create test conversation
-    const [conversation] = await db.insert(conversations).values({
-      id: randomUUID(),
-      userId: user.id,
-      title: "Test Conversation",
-    }).returning();
-
-    logger.info({ conversationId: conversation.id }, "Created test conversation");
-
-    // Create test messages
-    await db.insert(messages).values([
-      {
-        id: randomUUID(),
-        conversationId: conversation.id,
-        role: "user",
-        content: "Hello, AI!",
-      },
-      {
-        id: randomUUID(),
-        conversationId: conversation.id,
-        role: "assistant",
-        content: "Hello! How can I help you today?",
-      },
-    ]);
-
-    logger.info("Created test messages");
-    logger.info("Seeding completed successfully");
+    logger.info({ userId: user.id }, 'Created test user');
+    logger.info('Seeding completed successfully');
   } catch (error) {
-    logger.error({ error }, "Seeding failed");
+    logger.error({ error }, 'Seeding failed');
     process.exit(1);
   }
 
   await pool.end();
 }
 
-main(); 
+main().catch(error => {
+  logger.error({ error }, 'Unhandled error in main');
+  process.exit(1);
+});
