@@ -1,121 +1,69 @@
-import js from "@eslint/js";
-import eslintConfigPrettier from "eslint-config-prettier";
-import tseslint from "@typescript-eslint/eslint-plugin";
-import tsParser from "@typescript-eslint/parser";
-import pluginReactHooks from "eslint-plugin-react-hooks";
-import pluginReact from "eslint-plugin-react";
-import globals from "globals";
-import pluginNext from "@next/eslint-plugin-next";
+import js from '@eslint/js';
+import eslintConfigPrettier from 'eslint-config-prettier';
+import tseslint from '@typescript-eslint/eslint-plugin';
+import tsParser from '@typescript-eslint/parser';
+import pluginReactHooks from 'eslint-plugin-react-hooks';
+import pluginReact from 'eslint-plugin-react';
+import globals from 'globals';
+import pluginNext from '@next/eslint-plugin-next';
+import baseConfig from './base.js';
 
 /** @type {import("eslint").Linter.FlatConfig[]} */
 export default [
+  ...baseConfig,
   {
-    ignores: [
-      "node_modules/**",
-      "dist/**",
-      ".next/**",
-      "coverage/**",
-      "*.config.*",
-    ],
-  },
-  js.configs.recommended,
-  {
-    files: ["**/*.{ts,tsx,js,jsx}"],
+    files: ['**/*.{ts,tsx,js,jsx}'],
     plugins: {
-      "@typescript-eslint": tseslint,
+      '@typescript-eslint': tseslint,
       react: pluginReact,
-      "react-hooks": pluginReactHooks,
-      "@next/next": pluginNext,
+      'react-hooks': pluginReactHooks,
+      '@next/next': pluginNext,
     },
     languageOptions: {
-      parser: tsParser,
       parserOptions: {
-        project: true,
         ecmaFeatures: {
           jsx: true,
         },
       },
       globals: {
-        ...globals.browser,
-        ...globals.node,
-        ...globals.es2021,
-        React: "readonly",
+        React: 'readonly',
       },
     },
     settings: {
       react: {
-        version: "detect",
+        version: 'detect',
+      },
+      next: {
+        rootDir: ['apps/*/'],
       },
     },
     rules: {
-      ...eslintConfigPrettier.rules,
       ...pluginNext.configs.recommended.rules,
-      ...pluginNext.configs["core-web-vitals"].rules,
-      // TypeScript Rules
-      "@typescript-eslint/adjacent-overload-signatures": "error",
-      "@typescript-eslint/ban-types": "error",
-      "@typescript-eslint/no-explicit-any": "error",
-      "@typescript-eslint/no-unused-vars": [
-        "error",
-        {
-          argsIgnorePattern: "^_",
-          varsIgnorePattern: "^_",
-          caughtErrorsIgnorePattern: "^_",
-        },
-      ],
-      "@typescript-eslint/consistent-type-imports": [
-        "error",
-        { prefer: "type-imports", fixStyle: "inline-type-imports" },
-      ],
-      "@typescript-eslint/consistent-type-assertions": "error",
-      "@typescript-eslint/no-array-constructor": "error",
-      "@typescript-eslint/no-empty-interface": "error",
-      "@typescript-eslint/no-misused-new": "error",
-      "@typescript-eslint/no-namespace": "error",
-      "@typescript-eslint/no-this-alias": "error",
-      "@typescript-eslint/prefer-as-const": "error",
-      "@typescript-eslint/triple-slash-reference": "error",
-      "@typescript-eslint/no-non-null-assertion": "error",
-      "@typescript-eslint/no-inferrable-types": "error",
-      "@typescript-eslint/no-unnecessary-type-assertion": "error",
-      "@typescript-eslint/no-unnecessary-condition": "error",
-      "@typescript-eslint/strict-boolean-expressions": "error",
-      "@typescript-eslint/switch-exhaustiveness-check": "error",
-      "@typescript-eslint/prefer-optional-chain": "error",
-      "@typescript-eslint/prefer-nullish-coalescing": "error",
-      "@typescript-eslint/no-floating-promises": "error",
-      "@typescript-eslint/await-thenable": "error",
-      "@typescript-eslint/no-misused-promises": "error",
-      "@typescript-eslint/require-await": "error",
+      ...pluginNext.configs['core-web-vitals'].rules,
       // React Rules
-      "react/jsx-uses-react": "off",
-      "react/react-in-jsx-scope": "off",
-      "react/prop-types": "off",
-      "react/jsx-key": "error",
-      "react/jsx-no-duplicate-props": "error",
-      "react/jsx-no-undef": "error",
-      "react/no-children-prop": "error",
-      "react/no-danger-with-children": "error",
-      "react/no-deprecated": "error",
-      "react/no-direct-mutation-state": "error",
-      "react/no-find-dom-node": "error",
-      "react/no-is-mounted": "error",
-      "react/no-render-return-value": "error",
-      "react/no-string-refs": "error",
-      "react/no-unescaped-entities": "error",
-      "react/no-unknown-property": [
-        "error",
-        { ignore: ["cmdk-input-wrapper"] },
-      ],
-      "react/no-unsafe": "off",
+      'react/jsx-uses-react': 'off',
+      'react/react-in-jsx-scope': 'off',
+      'react/prop-types': 'off',
+      'react/jsx-key': 'error',
+      'react/jsx-no-duplicate-props': 'error',
+      'react/jsx-no-undef': 'error',
+      'react/no-children-prop': 'error',
+      'react/no-danger-with-children': 'error',
+      'react/no-deprecated': 'error',
+      'react/no-direct-mutation-state': 'error',
+      'react/no-find-dom-node': 'error',
+      'react/no-is-mounted': 'error',
+      'react/no-render-return-value': 'error',
+      'react/no-string-refs': 'error',
+      'react/no-unescaped-entities': 'error',
+      'react/no-unknown-property': ['error', { ignore: ['cmdk-input-wrapper'] }],
+      'react/no-unsafe': 'off',
       // React Hooks Rules
-      "react-hooks/rules-of-hooks": "error",
-      "react-hooks/exhaustive-deps": "error",
-      // General Rules
-      "no-console": ["error", { allow: ["warn", "error"] }],
-      "no-debugger": "error",
-      "no-alert": "error",
-      "no-unused-vars": "off", // Use @typescript-eslint/no-unused-vars instead
+      'react-hooks/rules-of-hooks': 'error',
+      'react-hooks/exhaustive-deps': 'error',
+      // Next.js specific rules
+      '@next/next/no-html-link-for-pages': ['error', 'apps/*/src/app'],
+      '@next/next/no-img-element': 'error',
     },
   },
 ];
