@@ -1,3 +1,3 @@
-import baseConfig from "@danky/eslint-config/base";
+import { baseConfig } from '@danky/eslint-config';
 
-export default [...baseConfig];
+export default baseConfig;
