@@ -9,39 +9,40 @@ function validateEnvironmentVariables(config: Config): void {
   const missingVars = new Set<string>();
 
   // Check LLM API keys
-  if (config.llm?.apiKey?.includes('${')) {
+  if (config.llm.apiKey !== undefined && config.llm.apiKey.includes('${')) {
     const matches = config.llm.apiKey.match(/\$\{([^}]+)\}/g);
-    if (matches) {
-      matches.forEach(match => {
-        const varName = match.slice(2, -1);
-        if (!process.env[varName]) {
-          missingVars.add(varName);
+    if (matches !== null) {
+      for (const match of matches) {
+        const envVar = match.slice(2, -1);
+        if (process.env[envVar] === undefined) {
+          throw new Error(`Missing environment variable: ${envVar}`);
         }
-      });
+      }
     }
   }
 
   // Check server environment variables
-  Object.entries(config.mcpServers || {}).forEach(([_, server]) => {
-    if (server.env) {
-      Object.values(server.env).forEach(value => {
-        if (typeof value === 'string' && value.includes('${')) {
-          const matches = value.match(/\$\{([^}]+)\}/g);
-          if (matches) {
-            matches.forEach(match => {
-              const varName = match.slice(2, -1);
-              if (!process.env[varName]) {
-                missingVars.add(varName);
-              }
-            });
+  for (const [, server] of Object.entries(config.mcpServers)) {
+    const serverEnv = server.env ?? {};
+    for (const value of Object.values(serverEnv)) {
+      if (value.includes('${')) {
+        const matches = value.match(/\$\{([^}]+)\}/g);
+        if (matches !== null) {
+          for (const match of matches) {
+            const varName = match.slice(2, -1);
+            if (process.env[varName] === undefined) {
+              missingVars.add(varName);
+            }
           }
         }
-      });
+      }
     }
-  });
+  }
 
   if (missingVars.size > 0) {
-    throw new Error(`Missing required environment variables: ${Array.from(missingVars).join(', ')}. Make sure these are set in the root .env file.`);
+    throw new Error(
+      `Missing required environment variables: ${Array.from(missingVars).join(', ')}. Make sure these are set in the root .env file.`
+    );
   }
 }
 
@@ -64,35 +65,38 @@ export interface Config {
   llm: LLMConfig;
   mcpServers: {
     [key: string]: MCPServerConfig;
-  }
+  };
 }
 
 export function loadConfig(configPath: string): Config {
+  let fileContents: string;
   try {
-    // Check if file exists
-    try {
-      readFileSync(configPath);
-    } catch (error) {
-      throw new Error(`Config file not found at "${configPath}". Error: ${error instanceof Error ? error.message : String(error)}`);
+    fileContents = readFileSync(configPath, 'utf-8');
+  } catch (error) {
+    throw new Error(
+      `Config file not found at "${configPath}". Error: ${error instanceof Error ? error.message : String(error)}`
+    );
+  }
+
+  // Replace environment variables in the format ${VAR_NAME} with their values
+  const configWithEnvVars = Object.entries(process.env).reduce((acc, [key, value]) => {
+    if (value !== undefined) {
+      const regex = new RegExp(`\\$\{${key}}`, 'g');
+      return acc.replace(regex, value);
     }
+    return acc;
+  }, fileContents);
 
-    let json5Str = readFileSync(configPath, 'utf-8');
+  let config: unknown;
+  try {
+    config = JSON5.parse(configWithEnvVars);
+  } catch (error) {
+    throw new Error(
+      `Failed to parse config file: ${error instanceof Error ? error.message : String(error)}`
+    );
+  }
 
-    // Replace environment variables in the format ${VAR_NAME} with their values
-    Object.entries(process.env).forEach(([key, value]) => {
-      const regex = new RegExp(`\\$\{${key}\}`, 'g');
-      if (value) {
-        json5Str = json5Str.replace(regex, value);
-      }
-    });
-
-    let config;
-    try {
-      config = JSON5.parse(json5Str);
-    } catch (error) {
-      throw new Error(`Failed to parse config file: ${error instanceof Error ? error.message : String(error)}`);
-    }
-
+  try {
     // Validate required fields
     validateConfig(config);
 
@@ -101,96 +105,114 @@ export function loadConfig(configPath: string): Config {
 
     return config;
   } catch (error) {
-    if (error instanceof Error) {
-      throw new Error(`Failed to load configuration from "${configPath}": ${error.message}`);
-    }
-    throw error;
+    throw new Error(
+      `Failed to load configuration from "${configPath}": ${error instanceof Error ? error.message : String(error)}`
+    );
   }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function validateConfig(config: any): asserts config is Config {
-  if (typeof config !== 'object' || config === null) {
+function validateConfig(config: unknown): asserts config is Config {
+  if (config === null || typeof config !== 'object') {
     throw new Error('Configuration must be an object');
   }
 
-  if (!config.llm) {
-    throw new Error('LLM configuration is required');
-  }
-  validateLLMConfig(config.llm);
+  const configObj = config as Record<string, unknown>;
 
-  if (typeof config.mcpServers !== 'object' || config.mcpServers === null) {
+  if (configObj.llm === null || typeof configObj.llm !== 'object') {
+    throw new Error('LLM configuration is required and must be an object');
+  }
+  validateLLMConfig(configObj.llm);
+
+  if (configObj.mcpServers === null || typeof configObj.mcpServers !== 'object') {
     throw new Error('mcpServers must be an object');
   }
 
-  Object.entries(config.mcpServers).forEach(([key, value]) => {
+  const servers = configObj.mcpServers as Record<string, unknown>;
+  for (const [key, value] of Object.entries(servers)) {
     try {
       validateMCPServerConfig(value);
     } catch (error) {
-      throw new Error(`Invalid configuration for MCP server "${key}": ${error instanceof Error ? error.message : String(error)}`);
+      throw new Error(
+        `Invalid configuration for MCP server "${key}": ${error instanceof Error ? error.message : String(error)}`
+      );
     }
-  });
+  }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function validateLLMConfig(llmConfig: any): asserts llmConfig is LLMConfig {
-  if (typeof llmConfig !== 'object' || llmConfig === null) {
+function validateLLMConfig(llmConfig: unknown): asserts llmConfig is LLMConfig {
+  if (llmConfig === null || typeof llmConfig !== 'object') {
     throw new Error('LLM configuration must be an object');
   }
 
-  if (typeof llmConfig.provider !== 'string') {
-    throw new Error('LLM provider must be a string');
+  const config = llmConfig as Record<string, unknown>;
+
+  if (typeof config.provider !== 'string' || config.provider === '') {
+    throw new Error('LLM provider must be a non-empty string');
   }
 
-  if (llmConfig.model !== undefined && typeof llmConfig.model !== 'string') {
-    throw new Error('LLM model must be a string if provided');
+  if (
+    'model' in config &&
+    (config.model === null || typeof config.model !== 'string' || config.model === '')
+  ) {
+    throw new Error('LLM model must be a non-empty string if provided');
   }
 
-  if (llmConfig.modelName !== undefined && typeof llmConfig.modelName !== 'string') {
-    throw new Error('LLM modelName must be a string if provided');
+  if (
+    'modelName' in config &&
+    (config.modelName === null || typeof config.modelName !== 'string' || config.modelName === '')
+  ) {
+    throw new Error('LLM modelName must be a non-empty string if provided');
   }
 
-  if (llmConfig.temperature !== undefined && typeof llmConfig.temperature !== 'number') {
+  if (
+    'temperature' in config &&
+    (config.temperature === null || typeof config.temperature !== 'number')
+  ) {
     throw new Error('LLM temperature must be a number if provided');
   }
 
-  if (llmConfig.maxTokens !== undefined && typeof llmConfig.maxTokens !== 'number') {
+  if (
+    'maxTokens' in config &&
+    (config.maxTokens === null || typeof config.maxTokens !== 'number')
+  ) {
     throw new Error('LLM maxTokens must be a number if provided');
   }
 
-  if (llmConfig.apiKey !== undefined && typeof llmConfig.apiKey !== 'string') {
-    throw new Error('LLM apiKey must be a string if provided');
+  if (
+    'apiKey' in config &&
+    (config.apiKey === null || typeof config.apiKey !== 'string' || config.apiKey === '')
+  ) {
+    throw new Error('LLM apiKey must be a non-empty string if provided');
   }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function validateMCPServerConfig(serverConfig: any): asserts serverConfig is MCPServerConfig {
-  if (typeof serverConfig !== 'object' || serverConfig === null) {
+function validateMCPServerConfig(serverConfig: unknown): asserts serverConfig is MCPServerConfig {
+  if (serverConfig === null || typeof serverConfig !== 'object') {
     throw new Error('MCP server configuration must be an object');
   }
 
-  if (typeof serverConfig.command !== 'string') {
-    throw new Error('MCP server command must be a string');
+  const config = serverConfig as Record<string, unknown>;
+
+  if (typeof config.command !== 'string' || config.command === '') {
+    throw new Error('MCP server command must be a non-empty string');
   }
 
-  if (!Array.isArray(serverConfig.args)) {
+  if (!Array.isArray(config.args)) {
     throw new Error('MCP server args must be an array');
   }
 
-  if (serverConfig.args.some((arg: unknown) => typeof arg !== 'string')) {
-    throw new Error('All MCP server args must be strings');
+  if (!config.args.every(arg => typeof arg === 'string' && arg !== '')) {
+    throw new Error('All MCP server args must be non-empty strings');
   }
 
-  if (serverConfig.env !== undefined) {
-    if (typeof serverConfig.env !== 'object' || serverConfig.env === null) {
+  if ('env' in config) {
+    if (config.env === null || typeof config.env !== 'object') {
       throw new Error('MCP server env must be an object if provided');
     }
 
-    // Validate that all env values are strings
-    for (const [, value] of Object.entries(serverConfig.env)) {
-      if (typeof value !== 'string') {
-        throw new Error('All MCP server env values must be strings');
-      }
+    const env = config.env as Record<string, unknown>;
+    if (!Object.values(env).every(value => typeof value === 'string' && value !== '')) {
+      throw new Error('All MCP server env values must be non-empty strings');
     }
   }
 }
