@@ -3,38 +3,49 @@
 
 type LogLevelString = 'trace' | 'debug' | 'info' | 'warn' | 'error' | 'fatal';
 
-enum LogLevel {
-  TRACE = 0,
-  DEBUG = 1,
-  INFO = 2,
-  WARN = 3,
-  ERROR = 4,
-  FATAL = 5
-}
+const LOG_LEVEL_VALUES = {
+  TRACE: 0,
+  DEBUG: 1,
+  INFO: 2,
+  WARN: 3,
+  ERROR: 4,
+  FATAL: 5,
+} as const;
+
+type LogLevel = (typeof LOG_LEVEL_VALUES)[keyof typeof LOG_LEVEL_VALUES];
 
 const LOG_COLORS = {
-  [LogLevel.TRACE]: '\x1b[90m',  // Gray
-  [LogLevel.DEBUG]: '\x1b[90m',  // Gray
-  [LogLevel.INFO]: '\x1b[90m',  // Gray
-  [LogLevel.WARN]: '\x1b[1;93m',  // Bold bright yellow
-  [LogLevel.ERROR]: '\x1b[1;91m',  // Bold bright red
-  [LogLevel.FATAL]: '\x1b[1;101m', // Red background, Bold text
+  [LOG_LEVEL_VALUES.TRACE]: '\x1b[90m', // Gray
+  [LOG_LEVEL_VALUES.DEBUG]: '\x1b[90m', // Gray
+  [LOG_LEVEL_VALUES.INFO]: '\x1b[90m', // Gray
+  [LOG_LEVEL_VALUES.WARN]: '\x1b[1;93m', // Bold bright yellow
+  [LOG_LEVEL_VALUES.ERROR]: '\x1b[1;91m', // Bold bright red
+  [LOG_LEVEL_VALUES.FATAL]: '\x1b[1;101m', // Red background, Bold text
 } as const;
 
 const LOG_LEVEL_MAP: Record<LogLevelString, LogLevel> = {
-  trace: LogLevel.TRACE,
-  debug: LogLevel.DEBUG,
-  info: LogLevel.INFO,
-  warn: LogLevel.WARN,
-  error: LogLevel.ERROR,
-  fatal: LogLevel.FATAL
+  trace: LOG_LEVEL_VALUES.TRACE,
+  debug: LOG_LEVEL_VALUES.DEBUG,
+  info: LOG_LEVEL_VALUES.INFO,
+  warn: LOG_LEVEL_VALUES.WARN,
+  error: LOG_LEVEL_VALUES.ERROR,
+  fatal: LOG_LEVEL_VALUES.FATAL,
+} as const;
+
+const LOG_LEVEL_NAMES: Record<LogLevel, string> = {
+  [LOG_LEVEL_VALUES.TRACE]: 'trace',
+  [LOG_LEVEL_VALUES.DEBUG]: 'debug',
+  [LOG_LEVEL_VALUES.INFO]: 'info',
+  [LOG_LEVEL_VALUES.WARN]: 'warn',
+  [LOG_LEVEL_VALUES.ERROR]: 'error',
+  [LOG_LEVEL_VALUES.FATAL]: 'fatal',
 } as const;
 
 export class Logger {
   private readonly level: LogLevel;
   private static readonly RESET = '\x1b[0m';
 
-  constructor({ level = LogLevel.INFO }: { level?: LogLevelString | LogLevel } = {}) {
+  constructor({ level = LOG_LEVEL_VALUES.INFO }: { level?: LogLevelString | LogLevel } = {}) {
     this.level = this.parseLogLevel(level);
   }
 
@@ -47,9 +58,18 @@ export class Logger {
     if (level < this.level) return;
 
     const color = LOG_COLORS[level];
-    const levelStr = `[${LogLevel[level].toLowerCase()}]`;
+    const levelStr = `[${LOG_LEVEL_NAMES[level]}]`;
+    const formattedArgs = args.map(this.formatValue);
 
-    console.log(`${color}${levelStr}${Logger.RESET}`, ...args.map(this.formatValue));
+    if (level >= LOG_LEVEL_VALUES.WARN) {
+      // Use console.warn or console.error for appropriate levels
+      const logMethod = level >= LOG_LEVEL_VALUES.ERROR ? console.error : console.warn;
+      logMethod(`${color}${levelStr}${Logger.RESET}`, ...formattedArgs);
+    } else {
+      // Use process.stdout.write for other levels to avoid console.log
+      const output = [`${color}${levelStr}${Logger.RESET}`, ...formattedArgs].join(' ') + '\n';
+      process.stdout.write(output);
+    }
   }
 
   private formatValue(value: unknown): string {
@@ -62,12 +82,12 @@ export class Logger {
     return (...args: unknown[]) => this.log(level, ...args);
   }
 
-  trace = this.createLogMethod(LogLevel.TRACE);
-  debug = this.createLogMethod(LogLevel.DEBUG);
-  info = this.createLogMethod(LogLevel.INFO);
-  warn = this.createLogMethod(LogLevel.WARN);
-  error = this.createLogMethod(LogLevel.ERROR);
-  fatal = this.createLogMethod(LogLevel.FATAL);
+  trace = this.createLogMethod(LOG_LEVEL_VALUES.TRACE);
+  debug = this.createLogMethod(LOG_LEVEL_VALUES.DEBUG);
+  info = this.createLogMethod(LOG_LEVEL_VALUES.INFO);
+  warn = this.createLogMethod(LOG_LEVEL_VALUES.WARN);
+  error = this.createLogMethod(LOG_LEVEL_VALUES.ERROR);
+  fatal = this.createLogMethod(LOG_LEVEL_VALUES.FATAL);
 }
 
 export interface LoggerOptions {
@@ -78,5 +98,5 @@ export function createLogger(options: LoggerOptions = {}): Logger {
   return new Logger(options);
 }
 
-export { LogLevel };
+export { LOG_LEVEL_VALUES as LogLevel };
 export type { LogLevelString };
