@@ -1,61 +1,7 @@
-// Components
-export * from './components/accordion'
-export * from './components/alert'
-export * from './components/alert-dialog'
-export * from './components/aspect-ratio'
-export * from './components/avatar'
-export * from './components/badge'
-export * from './components/button'
-export * from './components/calendar'
-export * from './components/card'
-export * from './components/carousel'
-export * from './components/checkbox'
-export * from './components/collapsible'
-export * from './components/command'
-export * from './components/context-menu'
-export * from './components/dialog'
-export * from './components/dropdown-menu'
-export * from './components/form'
-export * from './components/hover-card'
-export * from './components/icons'
-export * from './components/input'
-export * from './components/label'
-export * from './components/menubar'
-export * from './components/navigation-menu'
-export * from './components/popover'
-export * from './components/progress'
-export * from './components/radio-group'
-export * from './components/scroll-area'
-export * from './components/select'
-export * from './components/separator'
-export * from './components/sheet'
-export * from './components/skeleton'
-export * from './components/slider'
-export * from './components/switch'
-export * from './components/table'
-export * from './components/tabs'
-export * from './components/textarea'
-export * from './components/theme-provider'
-export * from './components/toast'
-export * from './components/toggle'
-export * from './components/tooltip'
-export * from './components/resizable'
-export * from './components/theme-toggle'
-export * from './components/loading-spinner'
-export * from './components/message-input'
+export { Input } from './components/input';
+export type { InputProps } from './components/input';
+export { Separator } from './components/separator';
+export type { SeparatorProps } from './components/separator';
 
-// Hooks
-export * from './hooks/use-toast'
-export * from './hooks/use-mobile'
-
-// Provider Components
-export * from './components/toaster'
-
-// Styles
-export * from './styles/fonts'
-
-// Utils
-export * from './lib/utils'
-
-// Types
-export type { ToastProps, ToastActionElement } from './components/toast'
+// Re-export utility functions
+export { cn } from './lib/utils';
